:art: Broke out components into seperate .js files
