+++ src/components/Game.js
@@ -1,0 +1,18 @@
+import React, { Component } from "react";
+import Board from "./Board";
+
+export default class Game extends Component {
+  render() {
+    return (
+      <div className="game">
+        <div className="game-board">
+          <Board />
+        </div>
+        <div className="game-info">
+          <div>{/* status */}</div>
+          <ol>{/* TODO */}</ol>
+        </div>
+      </div>
+    );
+  }
+}
