+++ src/components/Square.js
@@ -1,0 +1,14 @@
+import React, { Component } from "react";
+
+export default class Square extends Component {
+  render() {
+    return (
+      <button
+        className="square"
+        onClick={() => this.props.onClick({ value: "X" })}
+      >
+        {this.props.value}
+      </button>
+    );
+  }
+}
